+++ test/runtests.jl
@@ -105,7 +105,48 @@ end
         @test as(Symbol, "hello") === :hello
     end
 
-    @testset "to_same_conceret_type()" begin
+    @testset "Array axes and size" begin
+        @test TypeUtils.Dim === Int
+
+        @test as_array_dim(7) === 7
+        @test as_array_dim(Int16(9)) === 9
+        @test as_array_dim(Base.OneTo(5)) === 5
+        @test as_array_dim(-1:4) === 6
+        @test as_array_dim(Int8(0):Int8(3)) === 4
+
+        @test as_array_axis(7) === Base.OneTo(7)
+        @test as_array_axis(Int16(9)) === Base.OneTo(9)
+        @test as_array_axis(Base.OneTo(5)) === Base.OneTo(5)
+        @test as_array_axis(-1:4) === -1:4
+        @test as_array_axis(1:4) === 1:4
+        @test as_array_axis(Int8(0):Int8(3)) === 0:3
+
+        @test as_array_shape() === ()
+        @test as_array_shape(2,Int16(9),Int8(3)) === (2,9,3)
+        @test as_array_shape((2,Int16(9),Int8(3))) === (2,9,3)
+        @test as_array_shape(2,3,4) === (2,3,4)
+        @test as_array_shape(2,Base.OneTo{Int8}(3),4) === (2,3,4)
+        @test as_array_shape(2,Int8(1):Int8(3),4) === (Base.OneTo(2),1:3,Base.OneTo(4))
+        @test as_array_shape((0:2,-4:4,-2:1)) === (0:2,-4:4,-2:1)
+
+        @test as_array_axes() === ()
+        @test as_array_axes(2,Int16(9),Int8(3)) === (Base.OneTo(2),Base.OneTo(9),Base.OneTo(3))
+        @test as_array_axes((2,Int16(9),Int8(3))) === (Base.OneTo(2),Base.OneTo(9),Base.OneTo(3))
+        @test as_array_axes(2,3,4) === (Base.OneTo(2),Base.OneTo(3),Base.OneTo(4))
+        @test as_array_axes(2,Base.OneTo{Int8}(3),4) === (Base.OneTo(2),Base.OneTo(3),Base.OneTo(4))
+        @test as_array_axes(2,Int8(1):Int8(3),4) === (Base.OneTo(2),1:3,Base.OneTo(4))
+        @test as_array_axes((0:2,-4:4,-2:1)) === (0:2,-4:4,-2:1)
+
+        @test as_array_size() === ()
+        @test as_array_size(2,Int16(9),Int8(3)) === (2,9,3)
+        @test as_array_size((2,Int16(9),Int8(3))) === (2,9,3)
+        @test as_array_size(2,3,4) === (2,3,4)
+        @test as_array_size(2,Base.OneTo{Int8}(3),4) === (2,3,4)
+        @test as_array_size(2,Int8(1):Int8(3),4) === (2,3,4)
+        @test as_array_size((0:2,-4:4,-2:1)) === (3,9,4)
+    end
+
+    @testset "to_same_concrete_type()" begin
         @test to_same_concrete_type(Int) === Int
         @test to_same_concrete_type(UInt8, UInt8) === UInt8
         @test to_same_concrete_type(Int8, Int8, Int8) === Int8